strict-stm: split into submodules
Added missing:
* flushTBQueue
